refactor Encoding
