+++ test/Test/Asdf/EncodeSpec.hs
@@ -12,7 +12,6 @@ import Skeletest
 import Skeletest.Predicate qualified as P
 import Telescope.Asdf.Class
 import Telescope.Asdf.Core
-import Telescope.Asdf.Decoding
 import Telescope.Asdf.Encoding
 import Telescope.Asdf.Encoding.File
 import Telescope.Asdf.Error
